+++ src/de/omegazirkel/risingworld/tools/ui/PlayerPluginSettings.java
@@ -1,0 +1,9 @@
+package de.omegazirkel.risingworld.tools.ui;
+
+import net.risingworld.api.objects.Player;
+
+public abstract class PlayerPluginSettings {
+    public String pluginLabel;
+
+    public abstract BasePlayerPluginSettingsPanel createPlayerPluginSettingsUIElement(Player uiPlayer);
+}