+++ src/de/omegazirkel/risingworld/tools/ui/PluginMenuManager.java
@@ -3,11 +3,17 @@ package de.omegazirkel.risingworld.tools.ui;
 import java.util.ArrayList;
 import java.util.List;
 
+import de.omegazirkel.risingworld.OZTools;
+import de.omegazirkel.risingworld.tools.I18n;
 import net.risingworld.api.assets.TextureAsset;
 import net.risingworld.api.objects.Player;
 
 public class PluginMenuManager {
     private static List<MenuItem> menuItems = new ArrayList<>();
+
+    private static I18n t() {
+        return I18n.getInstance(OZTools.name);
+    };
 
     public static void registerPluginMenu(MenuItem menu) {
         menuItems.add(menu);
@@ -17,6 +23,21 @@ public class PluginMenuManager {
         // copy menuItems locally and add close button
         List<MenuItem> menuItemsCopy = new ArrayList<>(menuItems);
         menuItemsCopy.add(MenuItem.closeMenu(player));
+        menuItemsCopy
+                .add(new MenuItem(AssetManager.getIcon("icon-gpt-plugin-config"), t().get("TC_MENU_SETTINGS", player),
+                        (p) -> {
+                            p.hideRadialMenu(true);
+                            PlayerPluginSettingsOverlay overlay = (PlayerPluginSettingsOverlay) p
+                                    .getAttribute("tools.ui.overlay");
+                            if (overlay != null) {
+                                p.removeUIElement(overlay);
+                                p.deleteAttribute("tools.ui.overlay");
+                            }
+                            overlay = new PlayerPluginSettingsOverlay(p);
+                            CursorManager.show(p);
+                            p.addUIElement(overlay);
+                            p.setAttribute("tools.ui.overlay", overlay);
+                        }));
 
         showMenu(player, menuItemsCopy);
     }