+++ HISTORY.md
@@ -2,6 +2,10 @@
 
 ## [unreleased]
 
+## [0.14.0] - 2026-01-05 | Player-Plugin-Settings Manager
+
+- feat: new API for Plugin Settings, accessible through main radial menu
+
 ## [0.13.0] - 2025-12-28 | Refactor Consumer->Callback
 
 [BREAKING]